requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
